Refactored hateoas util classes
