+++ src/main/java/com/epam/esm/web/hateoas/ModelAssembler.java
@@ -4,5 +4,7 @@ import org.springframework.hateoas.PagedModel;
 import org.springframework.hateoas.server.SimpleRepresentationModelAssembler;
 
 public interface ModelAssembler<T> extends SimpleRepresentationModelAssembler<T> {
+
     void setMetadata(PagedModel.PageMetadata metadata);
+    void setModelLinkBuilder(ModelLinkBuilder<T> modelLinkBuilder);
 }
